Start Step2 ofArea Covering
* Fix a unit of XY from TwoGpsPositionToXY
* Add all the preliminary computations for Step2

+++ mission_manager/src/pixel_to_xy_gps_position.hpp
@@ -27,7 +27,7 @@
 #define PIXEL_TO_XY_GPS_POSITION_HPP
 
 double  HorizontalFOVLenght(double relative_altitude, double angle_fov_camera);
-double  VerticalFOVLenght(double hfov_lenght, int cam_width_pixel,  int cam_height_pixel);
+double  VerticalFOVLenght(double hfov_lenght, double cam_width_pixel,  double cam_height_pixel);
 void    TwoGpsPositionToXY(double latitude_1, double latitude_2, 
                             double longitude_1, double longitude_2, 
                             double & x, double & y);
